Updates to Mob Data
This commit changes a lot of mob data.
- New classes have been created for each Horse variant (Donkey,
Skeleton, Zombie, Mule).
- Rather than only scanning for the equivalent of adult/baby, the plugin
now detects whether or not "!adult" or "!baby" is contained anywhere in
the string.
- Rather than checking for saddled/unsaddled, the plugin now checks if
the term "!saddled" or "!unsaddled" is contained anywhere in the string.
- The horse class has gone through many changes to drop support for
variant, chest, and jump values. It now only s cans for the horse's
color, style, whether it is tamed, and the age. Please refer to the Wiki
(Creatures > Data Values > Horse) to find all available styles and
colors.
- The Zombie class has been changed to drop support for Villager Zombies

- To replace the above changes, a new Zombie Villager class had been
made.
For more information on this commit, please create an Issue / Help
Ticket!