Fixing Mob-Name Detection.
Previously an issue persisted with the detection of mob names (As
reported by Bukkit user Drexl_Levante). There have been quite a few code
changes to resolve this issue. To fix it, I did quite a bit of testing
in a developer environment. I've finally come up with a resolution. When
the drops file is parsed, the plugin will now detect when the customname
is not null (but is empty). It being empty was resulting in it not
getting sent back to the "matcher" Now the plugin will detect when it is
not null, but still is empty, and overwrite the empty name with
"CoolLordsWayToEnsureNobodyUsesThisNameHAHA"
- FOR SIMPLICITY PURPOSES, AND TO KEEP THE PLUGIN WORKING, PLEASE DO NOT
USE THIS NAME AS A REQUIREMENT -
Once the name has been overwritten, it gets sent to the match confirmer
(LivingEntityData). In the LED, there have been some overall code
cleanups. The main aspects that were changed were how the plugin
actually runs a match. Rather than detecting an empty name (seeing as
how that never got sent before!) the plugin now detects the
"CoolLordsWayToEnsureNobodyUsesThisNameHAHA" If the custom name is set
as that, it means you are essentially requiring a mob to not have a name
(specified by just a ~, ie. ZOMBIE~). If the name is meant to be clear,
it checks whether the mob has a name. If it does, the drop is denied
since the mob should NOT have a custom name. I've also split the checks
for if the mob's name is null, and if the two names don't match. In
either case, the drop is denied. Finally, I split up how the mob state
was seperated. It work's more efficiently now! To parse from the string,
it also ensures t he custom name is null and whether or not you have a
~. If it's null but a ~ is present, it rewrites customname.