Fixing how soil blocks work
This update fixes an issue reported by Discord user Yi. He noticed that
the plugin would illogically place (or rather, replace) soil and
soulsand blocks (which serve as farmland). The plugin would replace the
block itself if you wanted to place a crop on it. Ideally, it should
change the block above the farmland or soulsand into the crop or
netherwart, respectively. To correct this issue, I detected the material
being changed and if it matched up with either soulsand or farmland,
the plugin proceeds to detect what block you want to set it as. If you
require either a netherwart or one of the 6 crops, respectively, it will
modify the block being changed to the block directly above the
farmland/soulsand.