Configure update checker, restriction of enchants, and more!
Warning, this is a **PRERELEASE** meaning bugs may still occur. That is
why it is not published on Spigot! This commit contains a variety of
adjustments:
1. The config now has an update checker. Please regenerate the config by
deleting it (warning: all values will reset to their default! Make sure
you keep a backup!). If you wish to not delete it, simply add the values
`update_checker: true`  and `enchantments_restrict_matching: true`
Update Checker is pretty self explanatory. `Should we check for updates?
- set this to false if you don't want to be notified of new versions`
For the restrict matching, `Restrict matching: when true, it will
require enchantments to be exactly how they are configured in the drops
file (for example, A pickaxe defined with Fortune 3 will only let a
Fortune 3 pickaxe perform the drop. To allow a fortune 3, unbreaking 5,
and efficiency 5 pickaxe to perform the drop, set this value to false.`
Please message me on Discord if you need any help configuring this
(CoolLord22#7306).
2. In relation to the new config option, it is now used to determine
whether or not exact enchantments should be matched.
3. In relation to the new config option, the update checker will only be
run when the value is set to true.
4. Slight change on the updater (addition of else if rather than if) as
it was resulting in both messages being thrown.