Fix for Unbreaking enchantment not being accounted
This commit reworks how tool damage is done. If a tool has unbreaking on
it, the plugin will now detect the level of unbreaking it has and will
randomly generate an integer to calculate if it should do damage. It
uses the damage for tools (the chance to take damage is calculated by
100/level of unbreaking + 1), and compares the randomly generated
number with it. If the number is less than the chance of damage, it will
damage the tool. Otherwise, it will pass through without damaging.