Fixing an issue regarding the dropping of items
Issue reported by Discord User Folas1337. 
"Drops are not spawned like vanilla. When you let stone drop something else, then often times they spawn in a way which makes them bug "up" which means they work their way up until they find air. Just let stone drop something, go mining and when digging the blocks above you in a 1x1 tunnel, they'll very often appear above that block somewhere"
The issue has been addressed by changing the location to the exact center of the block. We do check first if the location returned is not already centered. This is checked by % 1, which would return 0 if the block was not centered. Since it returns as not centered, we're going to add 0.5 to that vector to center it! Thanks again to Folas1337 for reporting the bug.