Changes for vehicle optimization and item output
This update contains tons of changes, so please use it carefully. The
code has been redone to add support for all of the new vehicles that
weren't present in 1.8. In addition, the command /od id has been
modified to also output a line that will directly show the data of the
item you are holding. This will allow for easier transporting of item
information, by instructing the user exactly what item is being held
(including data values, enchantments, and lores).