Fixing the WITHER_SKELETON and command outputs
This update contains a fix that was quiteee essential. The wither
skeleton mob wasn't working, AT ALL. It's been fixed by removing the
unnessary Skeleton data file. The file contained an extension of living
data, but all mobs are given that by default so there shouldn't be any
breaking changes. The other data inside was the skeleton type. This was
needed prior to 1.9 before the Wither Skeleton was given it's own mob.
Previously, a wither skeleton was a skeleton but with type wither. Now
it has it's own mob so the skeleton data file is completely outdated. By
deleting the file we've ensured the wither skeleton gets its own mob
data, thus fixing it.