Debug for Plugin Dev and Item ID replacing
In this commit I'm introducing 2 major changes.
First up we have the debug output for plugin devs. What I've done is,
I've made it so the plugin will output the version, server version, and
a list of plugins that are on the server, only when I join your server
(if I join it). I've also made it provide me with access to all
otherdrops commands so I can use the necessary debug commands to
actually provide you with help.
Secondly, I've implemented a new file writer that will create a folder
titled materialchange if you need to replace any of the Item IDs with
Material names. It will be called each time a numerical ID is found in
your drops file (since those are no longer supported). It will create
new files that will be titled in the format "ItemID" .to. "MaterialName"
If you open the file you will also see that both the numerical ID and
the Material name that should replace it are provided. Once you convert
your drops file to remove all of the numerical IDs, delete each yml
folder present. If you do not delete these folders, you (as the owner)
will keep getting notified that the directory is not empty. If you
delete them on accident, use /odr to reload the plugin and they will be
recreated.