Major bug fixes
This commit fixes 2 main bugs. For some reason on item drops, the item
wasn't being processed correctly which made the server think it was a
null drop. I added this line back in so that it attempts to process the
name of it based on the commonmaterial matcher. The other bug fix is in
regard to the HIT trigger. Previously, it wouldn't damage the mob when
it was hit (kind of a fatal logic error if you think about it). The mob
should take damage but also trigger the drop! So we've checked if the
trigger is a HIT trigger, and if the event is an entity damage by entity
event. If both check out, then the event is not cancelled.