Support for WorldGuard, Towny, and GriefPrevention
This commit contains new support for WorldGuard, Towny, and
GriefPrevention. The plugin will now run a check to see whether the
player has permission to build in an area first. If the plugin detects
WorldGuard, Towny, or GriefPrevention it will run through each of the
respective checks. It will see if the player has been granted permission
from each of the respective plugins. If the player can build there, it
will run through the event. Otherwise it will cancel.