Modifying parsing of mob names
We are changing how mob names are parsed. The issue was reported by Discord user 521julian. The names would not parse properly if spaces were inside the config file. The plugin automatically replaced spaces with underscores, thus making the mob's name in-game different from the config. To fix this issue, we replace all underscores with spaces. However, this posed a new problem. The plugin would prevent underscores from actually being used. To bypass this issue, we introduced a new text replacement, only on creature names. If you desire an underscore in the mob's name, use "&_". This works by replacing the character sequence "&_" with the string "please do not use this text sequence it replaces an underscore" This allows for all remaining underscores to be replaced by a space character. Once the underscore parsing has been completed, the string is replaced by the character '_'