Modifying the parsing of material change
This commit contains a fix for a few issues that were happening due to
the material change addition. Now, we will check whether or not the
folder exists, and then proceed with the remaining checks. If the folder
doesn't exist, then there is no reason to try to see if it is empty!