Fixing the SKULL incorrectly dropping skulls
Previously the value of owner (which was taken from the config) was
being updated with the parsed variables. This meant if a user defined
"%v" for the Skull, it stopped trying to find the victim since %v was
overwritten with the first hit mob. To counter this, I created a
temporary storage variable that holds the value of the parsed owner
value, while leaving the original owner value alone. Another feature
added was with mobs. If you set a creature to drop their own head, it
will test to see whether an MHF_<creature> exists for the mob. If it
does, that head is used instead of the playername.