Fixing dispenser throwing NPE upon projectile hit
This commit fixes the null pointer exception thrown by a dispenser when
it hits a player with a projectile. This resulted because the shooter of
the projectile was actually null when we tried to parse it as an entity.
Since the dispenser is not an entity, we are returning null prematurely
to prevent the code from ever reaching the casting of an entity portion.
Fixes #24